requirements.txt
asgiref==3.4.1
click==8.0.1
Django==3.2.6
Flask==2.0.1
Flask-MySQLdb==0.2.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
Werkzeug==2.0.1
whitenoise==5.3.0
